Change ctrlp_custom_ignore to not ignore Directory.git
Consider a directory like `foo.git`.  Given that
this is a default name for a repo as shared by
Github, this is a fairly commmon case.

Within directories such as these, CtlrP will
return **NO ENTRIES**.  If you change the name
from `foo.git` to `foo_git` or just `foo`, you're
fine.

The Regex recommended by @kien in the CtrlP
documentation is:

```
\ 'dir':  '\v[\/]\.(git|hg|svn)$',
```

Using *this* pattern, CtlrP behaves as expected.